+++ lib/mongoid/relations/cascading/delete.rb
@@ -27,8 +27,18 @@ module Mongoid # :nodoc:
         #
         # @example Perform the cascading delete.
         #   strategy.cascade
+        #
+        # @since 2.0.0
         def cascade
-          Array.wrap(relation).each { |doc| doc.delete } if relation
+          if relation
+            if relation.cascades.empty?
+              safety = Threaded.safety_options
+              relation.clear
+              Threaded.safety_options = safety
+            else
+              ::Array.wrap(relation).each { |doc| doc.delete }
+            end
+          end
         end
       end
     end