+++ spec/app/models/home.rb
@@ -1,0 +1,4 @@
+class Home
+  include Mongoid::Document
+  belongs_to :person
+end