+++ spec/unit/mongoid/relations/cascading/delete_spec.rb
@@ -3,44 +3,98 @@ require "spec_helper"
 describe Mongoid::Relations::Cascading::Delete do
 
   let(:person) do
-    Person.new
-  end
-
-  let(:metadata) do
-    stub(:name => :posts)
-  end
-
-  let(:strategy) do
-    described_class.new(person, metadata)
+    Person.create
   end
 
   describe "#cascade" do
 
-    let(:post) do
-      stub
+    context "when cascading a has one" do
+
+      context "when the relation exists" do
+
+        let!(:home) do
+          person.create_home
+        end
+
+        before do
+          person.delete
+        end
+
+        it "deletes the relation" do
+          home.should be_destroyed
+        end
+
+        it "persists the deletion" do
+          expect {
+            home.reload
+          }.to raise_error(Mongoid::Errors::DocumentNotFound)
+        end
+      end
+
+      context "when the relation does not exist" do
+
+        before do
+          person.delete
+        end
+
+        it "deletes the base document" do
+          person.should be_destroyed
+        end
+      end
     end
+  end
 
-    context "when the documents exist" do
+  context "when cascading a has many" do
 
-      before do
-        person.expects(:posts).returns([ post ])
+    context "when the relation has documents" do
+
+      let!(:post_one) do
+        person.posts.create(:title => "one")
       end
 
-      it "deletes all documents in the relation" do
-        post.expects(:delete)
-        strategy.cascade
-      end
-    end
-
-    context "when no documents exist" do
-
-      before do
-        person.expects(:posts).returns([])
+      let!(:post_two) do
+        person.posts.create(:title => "two")
       end
 
-      it "does not delete anything" do
-        post.expects(:delete).never
-        strategy.cascade
+      context "when the documents are in memory" do
+
+        before do
+          post_one.expects(:delete).never
+          post_two.expects(:delete).never
+          person.delete
+        end
+
+        it "deletes the first document" do
+          post_one.should be_destroyed
+        end
+
+        it "deletes the second document" do
+          post_two.should be_destroyed
+        end
+
+        it "unbinds the first document" do
+          post_one.person.should be_nil
+        end
+
+        it "unbinds the second document" do
+          post_two.person.should be_nil
+        end
+
+        it "removes the documents from the relation" do
+          person.posts.should be_empty
+        end
+
+        it "persists the first deletion" do
+          expect {
+            post_one.reload
+          }.to raise_error(Mongoid::Errors::DocumentNotFound)
+        end
+
+        it "persists the second deletion" do
+          expect {
+            post_two.reload
+          }.to raise_error(Mongoid::Errors::DocumentNotFound)
+        end
       end
     end
   end